New architecture with segment_utils.py and segments.py added, and VDFSegmentGenerator removed from vdf_generator.py and added as a method of VDFImage instead.
