+++ pyxem/generators/vdf_generator.py
@@ -16,19 +16,17 @@
 # You should have received a copy of the GNU General Public License
 # along with pyXem.  If not, see <http://www.gnu.org/licenses/>.
 
-"""VDF generator, VDF segment generator and associated tools.
+"""VDF generator and associated tools.
 
 """
-
-from pyxem.signals.vdf_image import VDFImage, VDFSegment
-from pyxem.signals.diffraction_vectors import DiffractionVectors
-from pyxem.utils.vdf_utils import (normalize_vdf, separate)
-from pyxem.signals import (transfer_signal_axes,
-                           transfer_navigation_axes_to_signal_axes)
+import numpy as np
 
 from hyperspy.api import roi
-from hyperspy.signals import Signal2D
-import numpy as np
+
+from pyxem.signals.vdf_image import VDFImage
+from pyxem.utils.vdf_utils import normalize_vdf
+from pyxem.signals import (transfer_signal_axes,
+                           transfer_navigation_axes_to_signal_axes)
 
 
 class VDFGenerator:
@@ -152,124 +150,3 @@ class VDFGenerator:
 
         return vdfim
 
-
-class VDFSegmentGenerator:
-    """Generates VDF segments for the given VDFImages with corresponding
-    vector (aperture) positions.
-
-    Parameters
-    ----------
-    vdfs : VDFImage
-        The VDF images to be segmented.
-    vectors: DiffractionVectors, optional
-        The vector positions corresponding to the VDF images.
-
-    """
-
-    def __init__(self, vdfs, *args, **kwargs):
-
-        self.vdf_images = vdfs
-        self.vectors = vdfs.vectors
-
-    def get_vdf_segments(self, min_distance=1, min_size=1,
-                         max_size=np.inf, max_number_of_grains=np.inf,
-                         marker_radius=1, threshold=False,
-                         exclude_border=False):
-        """Separate segments from each of the VDF images using
-        edge-detection by the Sobel transform and the watershed
-        segmentation method implemented in scikit-image [1,2]. Obtain a
-        VDFSegment, similar to VDFImage, but where each image is a
-        segment of a VDF and the vectors correspond to each segment and
-        are not necessarily unique.
-
-        Parameters
-        ----------
-        min_distance: int
-            Minimum distance (in pixels) between grains required for
-            them to be considered as separate grains.
-        min_size : float
-            Grains with size (i.e. total number of pixels) below
-            min_size are discarded.
-        max_size : float
-            Grains with size (i.e. total number of pixels) above
-            max_size are discarded.
-        max_number_of_grains : int
-            Maximum number of grains included in the returned separated
-            grains. If it is exceeded, those with highest peak
-            intensities will be returned.
-        marker_radius : float
-            If 1 or larger, each marker for watershed is expanded to a disk
-            of radius marker_radius. marker_radius should not exceed
-            2*min_distance.
-        threshold: bool
-            If True, a mask is calculated by thresholding the VDF image
-            by the Li threshold method in scikit-image. If False
-            (default), the mask is the boolean VDF image.
-        exclude_border : int or True, optional
-            If non-zero integer, peaks within a distance of
-            exclude_border from the boarder will be discarded. If True,
-            peaks at or closer than min_distance of the boarder, will be
-            discarded.
-
-        References
-        ----------
-        [1] http://scikit-image.org/docs/dev/auto_examples/segmentation/
-            plot_watershed.html
-        [2] http://scikit-image.org/docs/dev/auto_examples/xx_applications/
-            plot_coins_segmentation.html#sphx-glr-auto-examples-xx-
-            applications-plot-coins-segmentation-py
-
-        Returns
-        -------
-        vdfsegs : VDFSegment
-            VDFSegment object containing segments (i.e. grains) of
-            single virtual dark field images with corresponding vectors.
-        """
-        vdfs = self.vdf_images
-        vectors = self.vectors.data
-
-        #TODO : Add aperture radius as an attribute of VDFImage and VDFSegment?
-
-        # Create an array of length equal to the number of vectors where each
-        # element is a np.object with shape (n: number of segments for this
-        # VDFImage, VDFImage size x, VDFImage size y).
-        vdfsegs = np.array(vdfs.map(
-            separate, show_progressbar=True, inplace=False,
-            min_distance=min_distance, min_size=min_size, max_size=max_size,
-            max_number_of_grains=max_number_of_grains,
-            marker_radius=marker_radius, threshold=threshold,
-            exclude_border=exclude_border), dtype=np.object)
-
-        segments, vectors_of_segments = [], []
-        for i, vector in zip(np.arange(vectors.size), vectors):
-            segments = np.append(segments, vdfsegs[i])
-            num_segs = np.shape(vdfsegs[i])[0]
-            vectors_of_segments = np.append(
-                vectors_of_segments, np.broadcast_to(vector, (num_segs, 2)))
-
-        if segments.dtype == object:
-            segments = segments.astype(float)
-
-        vectors_of_segments = vectors_of_segments.reshape((-1, 2))
-        segments = segments.reshape((np.shape(vectors_of_segments)[0],
-                                     vdfs.axes_manager.signal_shape[0],
-                                     vdfs.axes_manager.signal_shape[1]))
-
-        segments = Signal2D(segments).transpose(navigation_axes=[0],
-                                                signal_axes=[2, 1])
-
-        # Create VDFSegment and transfer axes calibrations
-        vdfsegs = VDFSegment(segments, DiffractionVectors(vectors_of_segments))
-        vdfsegs.segments = transfer_signal_axes(vdfsegs.segments,
-                                                self.vdf_images)
-        n = vdfsegs.segments.axes_manager.navigation_axes[0]
-        n.name = 'n'
-        n.units = 'number'
-        vdfsegs.vectors_of_segments.axes_manager.set_signal_dimension(1)
-        vdfsegs.vectors_of_segments = transfer_signal_axes(
-            vdfsegs.vectors_of_segments, self.vectors)
-        n = vdfsegs.vectors_of_segments.axes_manager.navigation_axes[0]
-        n.name = 'n'
-        n.units = 'number'
-
-        return vdfsegs
